requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
hdbscan>=0.8.28
matplotlib>=3.5.0
seaborn>=0.11.0
nltk>=3.7
langchain>=0.0.350
openai>=1.0.0
python-dotenv>=0.19.0
pathlib2>=2.3.0; python_version<"3.4"
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0 